requirements.txt
Flask==3.0.3
gunicorn==22.0.0
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
joblib==1.3.2
pytesseract==0.3.10
Pillow==10.3.0
nltk==3.8.1 